k8s: Table[*Endpoints]
