+++ pkg/k8s/resource_ctors.go
@@ -14,6 +14,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/cilium/cilium/pkg/hive/cell"
+	"github.com/cilium/cilium/pkg/hive/job"
 	cilium_api_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	cilium_api_v2alpha1 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2alpha1"
 	"github.com/cilium/cilium/pkg/k8s/client"
@@ -26,6 +27,10 @@ import (
 	"github.com/cilium/cilium/pkg/k8s/utils"
 	"github.com/cilium/cilium/pkg/k8s/version"
 	"github.com/cilium/cilium/pkg/node"
+	"github.com/cilium/cilium/pkg/statedb"
+	"github.com/cilium/cilium/pkg/statedb/index"
+	"github.com/cilium/cilium/pkg/statedb/reflector"
+	"github.com/cilium/cilium/pkg/time"
 )
 
 // Config defines the configuration options for k8s resources.
@@ -452,4 +457,129 @@ func CiliumClusterwideEnvoyConfigResource(lc cell.Lifecycle, cs client.Clientset
 		opts...,
 	)
 	return resource.New[*cilium_api_v2.CiliumClusterwideEnvoyConfig](lc, lw, resource.WithMetric("CiliumClusterwideEnvoyConfig")), nil
+}
+
+// ------
+// Resource[T] via Table[T] utilities
+
+type runtimeObjectWithName interface {
+	k8sRuntime.Object
+	GetName() string
+	GetNamespace() string
+}
+
+func newNameIndex[Obj runtimeObjectWithName]() statedb.Index[Obj, string] {
+	return statedb.Index[Obj, string]{
+		Name: "name",
+		FromObject: func(obj Obj) index.KeySet {
+			return index.NewKeySet(index.String(obj.GetNamespace() + "/" + obj.GetName()))
+		},
+		FromKey: index.String,
+		Unique:  true,
+	}
+}
+
+type tableResourceParams struct {
+	cell.In
+
+	Lifecycle cell.Lifecycle
+	Jobs      job.Registry
+	Scope     cell.Scope
+	DB        *statedb.DB
+}
+
+func newTableResource[Obj runtimeObjectWithName](p tableResourceParams, tableName string, nameIndex statedb.Index[Obj, string], transform func(any) (Obj, error), lw cache.ListerWatcher) (statedb.Table[Obj], resource.Resource[Obj], error) {
+	table, err := statedb.NewTable[Obj](tableName, nameIndex)
+	if err != nil {
+		return nil, nil, err
+	}
+	if err := p.DB.RegisterTable(table); err != nil {
+		return nil, nil, err
+	}
+
+	cfg := reflector.KubernetesConfig[Obj]{
+		BufferSize:     100,
+		BufferWaitTime: time.Millisecond * 100,
+		ListerWatcher:  lw,
+		Table:          table,
+		Transform:      transform,
+	}
+	p.Lifecycle.Append(
+		reflector.Kubernetes[Obj](
+			reflector.KubernetesParams[Obj]{
+				Config: cfg,
+				Jobs:   p.Jobs,
+				Scope:  p.Scope,
+				DB:     p.DB,
+			},
+		),
+	)
+
+	r := reflector.NewResourceAdapter[Obj](
+		p.DB,
+		table,
+		nameIndex,
+		func(key resource.Key) statedb.Query[Obj] {
+			return nameIndex.Query(key.Namespace + "/" + key.Name)
+		},
+	)
+
+	return table, r, nil
+}
+
+var ServiceNameIndex = newNameIndex[*slim_corev1.Service]()
+
+func NewServicesTableResource(p tableResourceParams, cfg Config, cs client.Clientset, opts ...func(*metav1.ListOptions)) (statedb.Table[*slim_corev1.Service], resource.Resource[*slim_corev1.Service], error) {
+	if !cs.IsEnabled() {
+		return nil, nil, nil
+	}
+
+	optsModifier, err := utils.GetServiceAndEndpointListOptionsModifier(cfg.K8sServiceProxyName)
+	if err != nil {
+		return nil, nil, err
+	}
+	lw := utils.ListerWatcherWithModifiers(
+		utils.ListerWatcherFromTyped[*slim_corev1.ServiceList](cs.Slim().CoreV1().Services("")),
+		append(opts, optsModifier)...,
+	)
+	return newTableResource[*slim_corev1.Service](p, "services", ServiceNameIndex, nil, lw)
+}
+
+var EndpointNameIndex = newNameIndex[*Endpoints]()
+
+func NewEndpointsTableResource(p tableResourceParams, cfg Config, cs client.Clientset) (statedb.Table[*Endpoints], resource.Resource[*Endpoints], error) {
+	if !cs.IsEnabled() {
+		return nil, nil, nil
+	}
+	endpointsOptsModifier, err := utils.GetServiceAndEndpointListOptionsModifier(cfg.K8sServiceProxyName)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	endpointSliceOpsModifier, err := utils.GetEndpointSliceListOptionsModifier()
+	if err != nil {
+		return nil, nil, err
+	}
+	lw := &endpointsListerWatcher{
+		cs:                         cs,
+		enableK8sEndpointSlice:     cfg.EnableK8sEndpointSlice,
+		endpointsOptsModifier:      endpointsOptsModifier,
+		endpointSlicesOptsModifier: endpointSliceOpsModifier,
+	}
+	transform := func(o any) (*Endpoints, error) {
+		out, err := transformEndpoint(o)
+		if err != nil {
+			return nil, err
+		}
+		return out.(*Endpoints), nil
+	}
+	return newTableResource[*Endpoints](p, "endpoints", EndpointNameIndex, transform, lw)
+	/*
+		return resource.New[*Endpoints](
+			lc,
+			lw,
+			resource.WithLazyTransform(lw.getSourceObj, transformEndpoint),
+			resource.WithMetric("Endpoint"),
+			resource.WithName("endpoints"),
+		), nil*/
 }
